move statistics state from Context to CountNodesVisitor
