+++ kernel/src/main/java/org/kframework/kil/loader/CountNodesVisitor.java
@@ -1,7 +1,6 @@
 // Copyright (c) 2013-2014 K Team. All Rights Reserved.
 package org.kframework.kil.loader;
 
-import org.kframework.kil.loader.Context;
 import org.kframework.kil.Cell;
 import org.kframework.kil.Configuration;
 import org.kframework.kil.Module;
@@ -10,19 +9,24 @@ import org.kframework.kil.Sentence;
 
 import org.kframework.kil.visitors.BasicVisitor;
 
+import java.util.Formatter;
+
 public class CountNodesVisitor extends BasicVisitor {
-    public CountNodesVisitor(Context context) {
-        super(context);
-        context.numModules = 0;
-        context.numSentences = 0;
-        context.numProductions = 0;
-        context.numCells = 0;
+
+    public int numModules, numSentences, numProductions, numCells;
+
+    public CountNodesVisitor() {
+        super(null);
+        numModules = 0;
+        numSentences = 0;
+        numProductions = 0;
+        numCells = 0;
     }
 
     @Override
     public Void visit(Module module, Void _) {
         if (!module.isPredefined()) {
-            context.numModules++;
+            numModules++;
             super.visit(module, _);
         }
         return null;
@@ -30,13 +34,13 @@ public class CountNodesVisitor extends BasicVisitor {
 
     @Override
     public Void visit(Sentence rule, Void _) {
-        context.numSentences++;
+        numSentences++;
         return super.visit(rule, _);
     }
 
     @Override
     public Void visit(Production production, Void _) {
-        context.numProductions++;
+        numProductions++;
         return super.visit(production, _);
     }
 
@@ -46,16 +50,25 @@ public class CountNodesVisitor extends BasicVisitor {
         inConfig = true;
         super.visit(config, _);
         inConfig = false;
-        context.numSentences++;
+        numSentences++;
         return null;
     }
 
     @Override
     public Void visit(Cell cell, Void _) {
         if (inConfig) {
-            context.numCells++;
+            numCells++;
             super.visit(cell, _);
         }
         return null;
+    }
+
+    public void printStatistics() {
+        Formatter f = new Formatter(System.out);
+        f.format("%n");
+        f.format("%-60s = %5d%n", "Number of Modules", numModules);
+        f.format("%-60s = %5d%n", "Number of Sentences", numSentences);
+        f.format("%-60s = %5d%n", "Number of Productions", numProductions);
+        f.format("%-60s = %5d%n", "Number of Cells", numCells);
     }
 }
